fixup query handling
Node.query() now supports a query and an optional in_args,
where either of those can be either an edn-formatted string
or a python object that will be converted to a string via
edn_format.dumps().

If the caller opts to pass the query as a python dictionary rather
than a string then it is incumbent on the caller to build that
dictionary using the edn_format Symbol and Keyword objects where
appropriate. See unit tests for examples.